🔨 Refactor imports to use files module for schema handling
